+++ examples/koa-ts/models/Movie.ts
@@ -1,4 +1,4 @@
-import { Model, JSONSchema } from 'objection'
+import { Model } from 'objection'
 import Person from './Person'
 
 export default class Movie extends Model {
@@ -13,7 +13,7 @@ export default class Movie extends Model {
   // Optional JSON schema. This is not the database schema! Nothing is generated
   // based on this. This is only used for validation. Whenever a model instance
   // is created it is checked against this schema. http://json-schema.org/.
-  static jsonSchema: JSONSchema = {
+  static jsonSchema = {
     type: 'object',
     required: ['name'],