+++ tests/integration/transactions.js
@@ -158,13 +158,13 @@ module.exports = session => {
         ]);
       })
         .then(() => {
-          return [
+          return Promise.all([
             session.knex('Model1').orderBy('model1Prop1'),
             session.knex('model2'),
             session.knex('Model1Model2')
-          ];
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(3);
           expect(_.map(rows1, 'model1Prop1')).to.eql(['a', 'b', 'd']);
           expect(rows2).to.have.length(1);
@@ -198,13 +198,13 @@ module.exports = session => {
           ]);
         })
         .then(() => {
-          return [
+          return Promise.all([
             session.knex('Model1').orderBy('model1Prop1'),
             session.knex('model2'),
             session.knex('Model1Model2')
-          ];
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(3);
           expect(_.map(rows1, 'model1Prop1')).to.eql(['a', 'b', 'd']);
           expect(rows2).to.have.length(1);
@@ -387,9 +387,13 @@ module.exports = session => {
         .catch(err => {
           expect(err.message).to.equal('whoops');
 
-          return [session.knex('Model1'), session.knex('model2'), session.knex('Model1Model2')];
+          return Promise.all([
+            session.knex('Model1'),
+            session.knex('model2'),
+            session.knex('Model1Model2')
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(0);
           expect(rows2).to.have.length(0);
           expect(rows3).to.have.length(0);
@@ -449,9 +453,13 @@ module.exports = session => {
         .catch(err => {
           expect(err.message).to.equal('whoops');
 
-          return [session.knex('Model1'), session.knex('model2'), session.knex('Model1Model2')];
+          return Promise.all([
+            session.knex('Model1'),
+            session.knex('model2'),
+            session.knex('Model1Model2')
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(0);
           expect(rows2).to.have.length(0);
           expect(rows3).to.have.length(0);
@@ -504,9 +512,13 @@ module.exports = session => {
         .catch(err => {
           expect(err.message).to.equal('whoops');
 
-          return [session.knex('Model1'), session.knex('model2'), session.knex('Model1Model2')];
+          return Promise.all([
+            session.knex('Model1'),
+            session.knex('model2'),
+            session.knex('Model1Model2')
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(0);
           expect(rows2).to.have.length(0);
           expect(rows3).to.have.length(0);
@@ -544,9 +556,13 @@ module.exports = session => {
         .catch(err => {
           expect(err.message).to.equal('whoops');
 
-          return [session.knex('Model1'), session.knex('model2'), session.knex('Model1Model2')];
+          return Promise.all([
+            session.knex('Model1'),
+            session.knex('model2'),
+            session.knex('Model1Model2')
+          ]);
         })
-        .spread((rows1, rows2, rows3) => {
+        .then(([rows1, rows2, rows3]) => {
           expect(rows1).to.have.length(0);
           expect(rows2).to.have.length(0);
           expect(rows3).to.have.length(0);