+++ tests/integration/upsertGraph.js
@@ -2,6 +2,7 @@ const expect = require('expect.js');
 const chai = require('chai');
 const Promise = require('bluebird');
 const { raw, transaction, ValidationError } = require('../../');
+const { createRejectionReflection } = require('../../testUtils/testUtils');
 const { FetchStrategy } = require('../../lib/queryBuilder/graph/GraphOptions');
 const mockKnexFactory = require('../../testUtils/mockKnex');
 
@@ -322,8 +323,8 @@ module.exports = session => {
                       ]
                     });
 
-                    return Promise.all([trx('Model1'), trx('model2')]).spread(
-                      (model1Rows, model2Rows) => {
+                    return Promise.all([trx('Model1'), trx('model2')]).then(
+                      ([model1Rows, model2Rows]) => {
                         // Row 5 should be deleted.
                         expect(model1Rows.find(it => it.id == 5)).to.equal(undefined);
                         // Row 6 should NOT be deleted even thought its parent is.
@@ -452,8 +453,8 @@ module.exports = session => {
                 ]
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 5 should be deleted.
                   expect(model1Rows.find(it => it.id == 5)).to.equal(undefined);
                   // Row 6 should NOT be deleted even thought its parent is.
@@ -795,8 +796,8 @@ module.exports = session => {
                   ]
                 });
 
-                return Promise.all([trx('Model1'), trx('model2')]).spread(
-                  (model1Rows, model2Rows) => {
+                return Promise.all([trx('Model1'), trx('model2')]).then(
+                  ([model1Rows, model2Rows]) => {
                     // Row 3 should NOT be deleted.
                     expect(model1Rows.find(it => it.id == 3)).to.eql({
                       id: 3,
@@ -1251,8 +1252,8 @@ module.exports = session => {
                 ]
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 3 should NOT be deleted.
                   expect(model1Rows.find(it => it.id == 3)).to.eql({
                     id: 3,
@@ -1381,8 +1382,8 @@ module.exports = session => {
                 ]
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 3 should NOT be deleted.
                   expect(model1Rows.find(it => it.id == 3)).to.eql({
                     id: 3,
@@ -1503,7 +1504,7 @@ module.exports = session => {
                 }
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(model1Rows => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(([model1Rows]) => {
                 // Row 3 should be deleted.
                 expect(model1Rows.find(it => it.id == 3)).to.equal(undefined);
               });
@@ -1542,8 +1543,8 @@ module.exports = session => {
                 }
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 3 should be deleted.
                   expect(model1Rows.find(it => it.id == 3)).to.equal(undefined);
                 }
@@ -1656,8 +1657,8 @@ module.exports = session => {
                 }
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 3 should not be deleted.
                   expect(model1Rows.find(it => it.id == 3)).to.not.equal(undefined);
                 }
@@ -2224,8 +2225,8 @@ module.exports = session => {
               ]
             });
 
-            return Promise.all([session.knex('Model1'), session.knex('model2')]).spread(
-              (model1Rows, model2Rows) => {
+            return Promise.all([session.knex('Model1'), session.knex('model2')]).then(
+              ([model1Rows, model2Rows]) => {
                 // Row 3 should NOT be deleted.
                 expect(model1Rows.find(it => it.id == 3)).to.eql({
                   id: 3,
@@ -2348,8 +2349,8 @@ module.exports = session => {
                 ]
               });
 
-              return Promise.all([trx('Model1'), trx('model2')]).spread(
-                (model1Rows, model2Rows) => {
+              return Promise.all([trx('Model1'), trx('model2')]).then(
+                ([model1Rows, model2Rows]) => {
                   // Row 5 should be deleted.
                   expect(model1Rows.find(it => it.id == 5)).to.equal(undefined);
                   // Row 6 should NOT be deleted even thought its parent is.
@@ -3501,7 +3502,7 @@ module.exports = session => {
           return Promise.map(fails, fail => {
             return transaction(session.knex, trx =>
               Model1.query(trx).upsertGraph(fail, { fetchStrategy })
-            ).reflect();
+            ).catch(err => createRejectionReflection(err));
           })
             .then(results => {
               // Check that all transactions have failed because of a validation error.
@@ -3568,8 +3569,8 @@ module.exports = session => {
                       ]
                     });
 
-                    return Promise.all([trx('Model1'), trx('model2')]).spread(
-                      (model1Rows, model2Rows) => {
+                    return Promise.all([trx('Model1'), trx('model2')]).then(
+                      ([model1Rows, model2Rows]) => {
                         // Row 5 should be deleted.
                         expect(model1Rows.find(it => it.id == 5)).to.equal(undefined);
                         // Row 6 should NOT be deleted even thought its parent is.
@@ -3809,7 +3810,7 @@ module.exports = session => {
           delete Model1.$$validator;
         });
 
-        it('should fail to do an incomplete upsert', () => {
+        it('should fail to do an incomplete upsert', async () => {
           const fails = [
             {
               id: 2,
@@ -3856,7 +3857,7 @@ module.exports = session => {
           return Promise.map(fails, fail => {
             return transaction(session.knex, trx =>
               Model1.query(trx).upsertGraph(fail, { update: true, fetchStrategy })
-            ).reflect();
+            ).catch(err => createRejectionReflection(err));
           })
             .then(results => {
               // Check that all transactions have failed because of a validation error.