+++ typings/objection/index.d.ts
@@ -95,12 +95,8 @@ declare namespace Objection {
   }
 
   export interface MixinFunction {
-    // Using ModelClass<M> causes TS 2.5 to render ModelClass<any> rather
-    // than an identity function type. <M extends typeof Model> retains the
-    // model subclass type in the return value, without requiring the user
-    // to type the Mixin call.
-    <MC extends ModelClass<any>>(modelClass: MC, ...plugins: Plugin[]): MC;
-    <MC extends ModelClass<any>>(modelClass: MC, plugins: Plugin[]): MC;
+    <MC extends AnyModelConstructor>(modelClass: MC, ...plugins: Plugin[]): MC;
+    <MC extends AnyModelConstructor>(modelClass: MC, plugins: Plugin[]): MC;
   }
 
   interface Aliasable {
@@ -159,7 +155,7 @@ declare namespace Objection {
 
   type Identity<T> = (value: T) => T;
   type AnyQueryBuilder = QueryBuilder<any, any>;
-  type AnyModelClass = ModelClass<any>;
+  type AnyModelConstructor = ModelConstructor<Model>;
   type ModifierFunction<QB extends AnyQueryBuilder> = (qb: QB, ...args: any[]) => void;
   type Modifier<QB extends AnyQueryBuilder = AnyQueryBuilder> =
     | ModifierFunction<QB>
@@ -325,7 +321,7 @@ declare namespace Objection {
   type PageQueryBuilder<QB extends AnyQueryBuilder> = QB['PageQueryBuilderType'];
 
   interface ForClassMethod {
-    <M extends Model>(modelClass: ModelClass<M>): QueryBuilderType<M>;
+    <M extends Model>(modelClass: ModelConstructor<M>): QueryBuilderType<M>;
   }
 
   /**
@@ -603,6 +599,14 @@ declare namespace Objection {
     (): boolean;
   }
 
+  interface HasMethod {
+    (selector: string | RegExp): boolean;
+  }
+
+  interface ClearMethod<QB extends AnyQueryBuilder> {
+    (selector: string | RegExp): QB;
+  }
+
   interface ColumnInfoMethod<QB extends AnyQueryBuilder> {
     (): Promise<knex.ColumnInfo>;
   }
@@ -612,11 +616,11 @@ declare namespace Objection {
   }
 
   interface AliasForMethod<QB extends AnyQueryBuilder> {
-    (modelClassOrTableName: string | ModelClass<any>, alias: string): QB;
+    (modelClassOrTableName: string | AnyModelConstructor, alias: string): QB;
   }
 
-  interface ModelClassMethod {
-    (): typeof Model;
+  interface ModelClassMethod<M extends Model> {
+    (): ModelClass<M>;
   }
 
   interface ReturningMethod {
@@ -779,6 +783,7 @@ declare namespace Objection {
 
   interface ModifiersMethod<QB extends AnyQueryBuilder> {
     (modifiers: Modifiers): QB;
+    (): QB;
   }
 
   // Deprecated
@@ -959,7 +964,7 @@ declare namespace Objection {
     findOne: WhereMethod<SingleQueryBuilder<this>>;
 
     execute(): Promise<R>;
-    castTo<MC extends Model>(modelClass: ModelClass<MC>): QueryBuilderType<MC>;
+    castTo<MC extends Model>(modelClass: ModelConstructor<MC>): QueryBuilderType<MC>;
 
     update(update: PartialModelObject<M>): NumberQueryBuilder<this>;
     update(): NumberQueryBuilder<this>;
@@ -1036,7 +1041,7 @@ declare namespace Objection {
     alias: OneArgMethod<string, this>;
     aliasFor: AliasForMethod<this>;
     withSchema: OneArgMethod<string, this>;
-    modelClass: ModelClassMethod;
+    modelClass: ModelClassMethod<M>;
     tableNameFor: TableRefForMethod;
     tableRefFor: TableRefForMethod;
     reject: OneArgMethod<any, this>;
@@ -1115,11 +1120,15 @@ declare namespace Objection {
     isDelete: BooleanReturningMethod;
     isRelate: BooleanReturningMethod;
     isUnrelate: BooleanReturningMethod;
+    isInternal: BooleanReturningMethod;
     hasWheres: BooleanReturningMethod;
     hasSelects: BooleanReturningMethod;
     // Deprecated
     hasEager: BooleanReturningMethod;
     hasWithGraph: BooleanReturningMethod;
+
+    has: HasMethod;
+    clear: ClearMethod<this>;
 
     clearSelect: IdentityMethod<this>;
     clearOrder: IdentityMethod<this>;
@@ -1171,8 +1180,8 @@ declare namespace Objection {
 
   export type RelationMappingsThunk = () => RelationMappings;
 
-  type ModelClassFactory = () => AnyModelClass;
-  type ModelClassSpecifier = ModelClassFactory | AnyModelClass | string;
+  type ModelClassFactory = () => AnyModelConstructor;
+  type ModelClassSpecifier = ModelClassFactory | AnyModelConstructor | string;
   type RelationMappingHook<M extends Model> = (
     model: M,
     context: QueryContext
@@ -1276,6 +1285,7 @@ declare namespace Objection {
   export interface SnakeCaseMappersOptions {
     upperCase?: boolean;
     underscoreBeforeDigits?: boolean;
+    underscoreBetweenUppercaseLetters?: boolean;
   }
 
   export interface ColumnNameMappers {
@@ -1356,7 +1366,133 @@ declare namespace Objection {
     new (): T;
   }
 
-  export interface ModelClass<M> extends Constructor<M> {}
+  export interface ModelConstructor<M extends Model> extends Constructor<M> {}
+
+  export interface ModelClass<M extends Model> extends ModelConstructor<M> {
+    QueryBuilder: typeof QueryBuilder;
+
+    tableName: string;
+    idColumn: string | string[];
+    jsonSchema: JSONSchema;
+    relationMappings: RelationMappings | RelationMappingsThunk;
+    modelPaths: string[];
+    jsonAttributes: string[];
+    virtualAttributes: string[];
+    uidProp: string;
+    uidRefProp: string;
+    dbRefProp: string;
+    propRefRegex: RegExp;
+    pickJsonSchemaProperties: boolean;
+    relatedFindQueryMutates: boolean;
+    relatedInsertQueryMutates: boolean;
+    modifiers: Modifiers;
+    columnNameMappers: ColumnNameMappers;
+
+    raw: RawFunction;
+    ref: ReferenceFunction;
+    fn: FunctionFunction;
+
+    BelongsToOneRelation: RelationType;
+    HasOneRelation: RelationType;
+    HasManyRelation: RelationType;
+    ManyToManyRelation: RelationType;
+    HasOneThroughRelation: RelationType;
+
+    defaultGraphOptions?: GraphOptions;
+    // Deprecated
+    defaultEagerAlgorithm?: EagerAlgorithm;
+    // Deprecated
+    defaultEagerOptions?: EagerOptions;
+
+    // Deprecated
+    WhereInEagerAlgorithm: EagerAlgorithm;
+    // Deprecated
+    NaiveEagerAlgorithm: EagerAlgorithm;
+    // Deprecated
+    JoinEagerAlgorithm: EagerAlgorithm;
+
+    query(this: Constructor<M>, trxOrKnex?: TransactionOrKnex): QueryBuilderType<M>;
+
+    relatedQuery<K extends keyof M>(
+      relationName: K,
+      trxOrKnex?: TransactionOrKnex
+    ): ArrayRelatedQueryBuilder<M[K]>;
+
+    relatedQuery<RM extends Model>(
+      relationName: string,
+      trxOrKnex?: TransactionOrKnex
+    ): QueryBuilderType<RM>;
+
+    fromJson(json: object, opt?: ModelOptions): M;
+    fromDatabaseJson(json: object): M;
+
+    createValidator(): Validator;
+    createValidationError(args: CreateValidationErrorArgs): Error;
+    createNotFoundError(): Error;
+
+    tableMetadata(opt?: TableMetadataOptions): TableMetadata;
+    fetchTableMetadata(opt?: FetchTableMetadataOptions): Promise<TableMetadata>;
+
+    knex(knex?: knex): knex;
+    knexQuery(): knex.QueryBuilder;
+    startTransaction(knexOrTransaction?: TransactionOrKnex): Promise<Transaction>;
+
+    transaction<T>(callback: (trx: Transaction) => Promise<T>): Promise<T>;
+    transaction<T>(
+      trxOrKnex: TransactionOrKnex,
+      callback: (trx: Transaction) => Promise<T>
+    ): Promise<T>;
+
+    bindKnex(trxOrKnex: TransactionOrKnex): this;
+    bindTransaction(trxOrKnex: TransactionOrKnex): this;
+
+    // Deprecated
+    loadRelated(
+      modelOrObject: PartialModelObject<M>,
+      expression: RelationExpression<M>,
+      modifiers?: Modifiers<QueryBuilderType<M>>,
+      trxOrKnex?: TransactionOrKnex
+    ): SingleQueryBuilder<QueryBuilderType<M>>;
+
+    // Deprecated
+    loadRelated(
+      modelOrObject: PartialModelObject<M>[],
+      expression: RelationExpression<M>,
+      modifiers?: Modifiers<QueryBuilderType<M>>,
+      trxOrKnex?: TransactionOrKnex
+    ): QueryBuilderType<M>;
+
+    fetchGraph(
+      modelOrObject: PartialModelObject<M>,
+      expression: RelationExpression<M>,
+      options?: FetchGraphOptions
+    ): SingleQueryBuilder<QueryBuilderType<M>>;
+
+    fetchGraph(
+      modelOrObject: PartialModelObject<M>[],
+      expression: RelationExpression<M>,
+      options?: FetchGraphOptions
+    ): QueryBuilderType<M>;
+
+    getRelations(): Relations;
+    getRelation(name: string): Relation;
+
+    traverse(models: Model | Model[], traverser: TraverserFunction): void;
+    traverse(
+      filterConstructor: ModelConstructor<Model>,
+      models: Model | Model[],
+      traverser: TraverserFunction
+    ): void;
+
+    beforeFind(args: StaticHookArguments<any>): any;
+    afterFind(args: StaticHookArguments<any>): any;
+    beforeInsert(args: StaticHookArguments<any>): any;
+    afterInsert(args: StaticHookArguments<any>): any;
+    beforeUpdate(args: StaticHookArguments<any>): any;
+    afterUpdate(args: StaticHookArguments<any>): any;
+    beforeDelete(args: StaticHookArguments<any>): any;
+    afterDelete(args: StaticHookArguments<any>): any;
+  }
 
   export class Model {
     static QueryBuilder: typeof QueryBuilder;
@@ -1402,12 +1538,12 @@ declare namespace Objection {
     static JoinEagerAlgorithm: EagerAlgorithm;
 
     static query<M extends Model>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       trxOrKnex?: TransactionOrKnex
     ): QueryBuilderType<M>;
 
     static relatedQuery<M extends Model, K extends keyof M>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       relationName: K,
       trxOrKnex?: TransactionOrKnex
     ): ArrayRelatedQueryBuilder<M[K]>;
@@ -1417,8 +1553,8 @@ declare namespace Objection {
       trxOrKnex?: TransactionOrKnex
     ): QueryBuilderType<RM>;
 
-    static fromJson<M extends Model>(this: ModelClass<M>, json: object, opt?: ModelOptions): M;
-    static fromDatabaseJson<M extends Model>(this: ModelClass<M>, json: object): M;
+    static fromJson<M extends Model>(this: Constructor<M>, json: object, opt?: ModelOptions): M;
+    static fromDatabaseJson<M extends Model>(this: Constructor<M>, json: object): M;
 
     static createValidator(): Validator;
     static createValidationError(args: CreateValidationErrorArgs): Error;
@@ -1429,7 +1565,7 @@ declare namespace Objection {
 
     static knex(knex?: knex): knex;
     static knexQuery(): knex.QueryBuilder;
-    static startTransaction(knexOrTransaction?: TransactionOrKnex): Transaction;
+    static startTransaction(knexOrTransaction?: TransactionOrKnex): Promise<Transaction>;
 
     static transaction<T>(callback: (trx: Transaction) => Promise<T>): Promise<T>;
     static transaction<T>(
@@ -1442,7 +1578,7 @@ declare namespace Objection {
 
     // Deprecated
     static loadRelated<M extends Model>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       modelOrObject: PartialModelObject<M>,
       expression: RelationExpression<M>,
       modifiers?: Modifiers<QueryBuilderType<M>>,
@@ -1451,7 +1587,7 @@ declare namespace Objection {
 
     // Deprecated
     static loadRelated<M extends Model>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       modelOrObject: PartialModelObject<M>[],
       expression: RelationExpression<M>,
       modifiers?: Modifiers<QueryBuilderType<M>>,
@@ -1459,14 +1595,14 @@ declare namespace Objection {
     ): QueryBuilderType<M>;
 
     static fetchGraph<M extends Model>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       modelOrObject: PartialModelObject<M>,
       expression: RelationExpression<M>,
       options?: FetchGraphOptions
     ): SingleQueryBuilder<QueryBuilderType<M>>;
 
     static fetchGraph<M extends Model>(
-      this: ModelClass<M>,
+      this: Constructor<M>,
       modelOrObject: PartialModelObject<M>[],
       expression: RelationExpression<M>,
       options?: FetchGraphOptions
@@ -1575,14 +1711,14 @@ declare namespace Objection {
    * @tutorial https://vincit.github.io/objection.js/guide/transactions.html#creating-a-transaction
    */
   export interface transaction {
-    start(knexOrModel: knex | ModelClass<any>): Promise<Transaction>;
+    start(knexOrModel: knex | AnyModelConstructor): Promise<Transaction>;
 
-    <MC1 extends ModelClass<any>, ReturnValue>(
+    <MC1 extends AnyModelConstructor, ReturnValue>(
       modelClass1: MC1,
       callback: (boundModelClass: MC1, trx?: Transaction) => Promise<ReturnValue>
     ): Promise<ReturnValue>;
 
-    <MC1 extends ModelClass<any>, MC2 extends ModelClass<any>, ReturnValue>(
+    <MC1 extends AnyModelConstructor, MC2 extends AnyModelConstructor, ReturnValue>(
       modelClass1: MC1,
       modelClass2: MC2,
       callback: (
@@ -1593,9 +1729,9 @@ declare namespace Objection {
     ): Promise<ReturnValue>;
 
     <
-      MC1 extends ModelClass<any>,
-      MC2 extends ModelClass<any>,
-      MC3 extends ModelClass<any>,
+      MC1 extends AnyModelConstructor,
+      MC2 extends AnyModelConstructor,
+      MC3 extends AnyModelConstructor,
       ReturnValue
     >(
       modelClass1: MC1,
@@ -1610,10 +1746,10 @@ declare namespace Objection {
     ): Promise<ReturnValue>;
 
     <
-      MC1 extends ModelClass<any>,
-      MC2 extends ModelClass<any>,
-      MC3 extends ModelClass<any>,
-      MC4 extends ModelClass<any>,
+      MC1 extends AnyModelConstructor,
+      MC2 extends AnyModelConstructor,
+      MC3 extends AnyModelConstructor,
+      MC4 extends AnyModelConstructor,
       ReturnValue
     >(
       modelClass1: MC1,
@@ -1630,11 +1766,11 @@ declare namespace Objection {
     ): Promise<ReturnValue>;
 
     <
-      MC1 extends ModelClass<any>,
-      MC2 extends ModelClass<any>,
-      MC3 extends ModelClass<any>,
-      MC4 extends ModelClass<any>,
-      MC5 extends ModelClass<any>,
+      MC1 extends AnyModelConstructor,
+      MC2 extends AnyModelConstructor,
+      MC3 extends AnyModelConstructor,
+      MC4 extends AnyModelConstructor,
+      MC5 extends AnyModelConstructor,
       ReturnValue
     >(
       modelClass1: MC1,
@@ -1658,8 +1794,8 @@ declare namespace Objection {
   }
 
   interface initialize {
-    (knex: Knex, modelClasses: ModelClass<any>[]): Promise<void>;
-    (modelClasses: ModelClass<any>[]): Promise<void>;
+    (knex: Knex, modelClasses: AnyModelConstructor[]): Promise<void>;
+    (modelClasses: AnyModelConstructor[]): Promise<void>;
   }
 
   /**
@@ -1688,7 +1824,9 @@ declare namespace Objection {
     | 'boolean'
     | 'object'
     | 'array'
-    | 'null';
+    | 'null'
+    | string;
+
   export type JSONSchemaType = JSONSchemaArray[] | boolean | number | null | object | string;
 
   // Workaround for infinite type recursion